Fix bug with optimize_hyperparams
The bug causes the last evaluated hyperparams to be returned as best,
regardless of fitness. Reusing the same dict for hyperparameters resulted
in it being modified, instead of a new hyperparameter dict being returned.
This was not an issue before Optimizer.optimize returned the result of
decode, because the bit string would previously decode into the correct
dictionary.